requirements.txt
pyairtable
python-dotenv
openai
requests
